+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - close single spider image auto download image
 - update single spider image code
 - add jm id download image
+- add detect jm domain code
 
 ## v1.1.7-beta.2.240617
 ### 2024/6/17